Update for >= Kerbalism 3.30 support
Kerbalism 3.30 removes the field "control_path" in favor of each connection only having a hop based "next_hop" bit of GUID info.   This is an easy correction, but not API compatible when using reflection as is done here.  So we add a version detect to Kerbalism and do a simple boolean set and if check based on the file version (which Kerbalism keeps up to date).  This enables compatability with both new and old builds of the mod.